// client/src/game/shared.ts
export type MapLocation = typeof MAPS[number];

export interface ILocation {
  pos: { x: number; y: number };
  map: MapLocation;
}

export interface ICreateAccount {
  username: string;
  password: string;
  dna: IUserDNA;
}

export type UserSprite = 'c1';
export type EnemySprite = 'e1';
// ...
export type AttackType = 'slash' | 'thrust' | 'spell' | 'oversize' | 'shoot';

export interface IEnemyAttacked {
  id: string;
  dead: boolean;
  damage: number;
  attributes: StrictMap<CombatAttribute, number>;
  loot: IItem[];
}

export interface IUserAttack {
  name: string;
  enemies: IEnemyAttacked[];
}

export interface IUpdateLoot {
  id: string;
  items: IItem[];
}

export const COMBAT_ATTRIBUTES = ['health', 'stamina', 'mana'] as const;
export type CombatAttribute = typeof COMBAT_ATTRIBUTES[number];

export const LIFE_ATTRIBUTES = ['hunger', 'exhaustion', 'age'] as const;
export type LifeAttribute = typeof LIFE_ATTRIBUTES[number];

export type UserAttribute = LifeAttribute | CombatAttribute;

export const COMBAT_STATS = [
  'health',
  'stamina',
  'mana',
  'agility',
  'defense',
  'accuracy',
  'strength',
  'intelligence',
  'dexterity',
  'sneak',
  'awareness',
  'slash',
  'thrust',
  'spell',
  'oversize',
  'shoot',
] as const;
export type CombatStat = typeof COMBAT_STATS[number];

export const GATHERING_STATS = [
  'woodcutting',
  'mining',
  'fishing',
  'farming',
] as const;
export type GatheringStat = typeof GATHERING_STATS[number];

export const ARTISAN_STATS = [
  'cooking',
  'crafting',
  'enchanting',
  'alchemy',
] as const;
export type ArtisanStat = typeof ARTISAN_STATS[number];

export const SUPPORT_STATS = ['thieving'] as const;
export type SupportStat = typeof SUPPORT_STATS[number];

export const USER_STATS = [
  ...COMBAT_STATS,
  ...GATHERING_STATS,
  ...ARTISAN_STATS,
  ...SUPPORT_STATS,
] as const;

export type UserStat = typeof USER_STATS[number];

export interface IStat {
  level: number;
  xp: number;
  nextXp: number;
  points: number;
}

const ITEM_TYPES = ['WEARABLE', 'USEABLE', 'CONSUMABLE'] as const;
type ItemType = typeof ITEM_TYPES[number];

export interface IEquipItem {
  name: string;
  equipement: Partial<IUserEquipement>;
  inventory?: IItem[];
}

export interface IItem {
  name: string;
  sprite: string;
  quantity: number;
  type: ItemType;
  sex?: 'male' | 'female';
  bonus?: Partial<StrictMap<UserStat, number>>;
  duration?: number;
  activateJob?: GatheringStat | ArtisanStat;
  equipSlot?: EquipementSlot;
  equipSprite?: string;
  attackType?: AttackType;
}

export const EQUIPEMENT_SLOTS = [
  'earring1',
  'head',
  'earring2',
  'shoulders',
  'torso',
  'cape',
  'necklace',
  'weapon',
  'arms',
  'belt',
  'hands',
  'shield',
  'ring1',
  'legs',
  'feet',
  'ring2',
] as const;

export type EquipementSlot = typeof EQUIPEMENT_SLOTS[number];
// ...
export interface IEnemy {
  id: string;
  location: ILocation;
  sprite: EnemySprite;
  name: string;
  stats: StrictMap<CombatStat, IStat>;
  attributes: StrictMap<CombatAttribute, number>;
  threat: number;
  tracked?: boolean;
  dead: boolean;
}

export const TASKS = ['routine', 'follow', 'dead'] as const;
export type TaskType = typeof TASKS[number];

export interface TrackAI {
  map: MapLocation;
  enemies: AIEnemy[];
}

export interface AIEnemy {
  id: string;
  movement: AIMovement;
  current: TaskType;
}

export interface AIMovement {
  velx: number;
  vely: number;
  dir: Direction;
  anim: PlayerAnimation;
  pos: { x: number; y: number };
}

export interface InitGame {
  map: MapLocation;
  users: IUser[];
  enemies: IEnemy[];
  loots: IUpdateLoot[];
}

export interface IPlayerMovement {
  name: string;
  dir: Direction | 'end';
  speed?: number;
  pos?: { x: number; y: number };
}

export const DIRECTIONS = ['up', 'down', 'left', 'right'] as const;
export type Direction = typeof DIRECTIONS[number];
// ...
export type PlayerAnimation =
  | PlayableAnimations
  | IdleDirection
  | MoveDirection;